requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
elevenlabs==1.50.3; python_version >= '3.8' and python_version < '4.0'
fastapi==0.115.6; python_version >= '3.8'
ffmpy==0.5.0; python_version >= '3.8' and python_version < '4.0'
filelock==3.16.1; python_version >= '3.8'
fsspec==2024.12.0; python_version >= '3.8'
gradio==4.19.2
gradio-client==1.5.4; python_version >= '3.10'
groq==0.15.0
gtts==2.5.4; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
huggingface-hub==0.27.1; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jinja2==3.1.5; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
numpy==2.2.1; python_version >= '3.10'
orjson==3.10.14; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
pyaudio==0.2.14
pydantic==2.10.5; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydub==0.25.1
pygments==2.19.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
requests==2.31.0
rich==13.9.4; python_full_version >= '3.8.0'
ruff==0.9.1; sys_platform != 'emscripten'
safehttpx==0.1.6; python_version >= '3.10'
semantic-version==2.10.0; python_version >= '2.7'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
speechrecognition==3.10.1
starlette==0.41.3; sys_platform != 'emscripten'
tomlkit==0.13.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typer==0.15.1; sys_platform != 'emscripten'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
uvicorn==0.34.0; sys_platform != 'emscripten'
websockets==14.1; python_version >= '3.9'
python-dotenv==1.0.1
streamlit==1.32.0
SpeechRecognition==3.10.1
Pillow==10.2.0
